Added obstacle edges detection when moving a sprite.
This permitted sprites to avoid corners when walking/running
and makes the game a bit smoother for the player.

This one went through a lot of testing to avoid breaking the current
game mechanisms.

Closes #208

+++ src/modes/map/map_objects.h
@@ -1034,10 +1034,10 @@ public:
     //! \param x x location on collision grid
     //! \param y y location on collision grid
     //! \return whether the location would be a "wall" for the party or not
-    bool IsStaticCollision(uint32 x, uint32 y);
+    bool IsStaticCollision(float x, float y);
 
     //! \brief checks if the location on the grid has a simple map collision. This is different from
-    //! IsStaticCollision, int hat it DOES NOT check static objects, but only the collision value for the map
+    //! IsStaticCollision, in that it DOES NOT check static objects, but only the collision value for the map
     bool IsMapCollision(uint32 x, uint32 y)
     { return (_collision_grid[y][x] > 0); }
